requirements.txt
asttokens==2.2.1
backcall==0.2.0
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fonttools==4.41.1
ipykernel==6.25.0
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
matplotlib==3.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.25.1
opencv-python==4.8.0.74
packaging==23.1
parso==0.8.3
pdf2image==1.16.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==25.1.0
six==1.16.0
stack-data==0.6.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
wcwidth==0.2.6
